Finished the last 12 quizzes.

+++ week-01/Solidity-Speed-Quiz/src/Bytes.sol
@@ -3,6 +3,7 @@ pragma solidity 0.8.28;
 
 contract Bytes {
     bytes b;
+
     function main(uint8 z) public returns (bytes memory) {
         // if z = 0, return empty bytes
         // if z = 1, return bytes 0x00
@@ -11,11 +12,10 @@ contract Bytes {
         // if z = 4, return bytes 0x00010203
         // ...
         // etc
-        
-        for(uint i=0; i<z; i++){
-           b = abi.encodePacked(i-1);  
+
+        for (uint256 i = 0; i < z; i++) {
+            b = abi.encodePacked(i - 1);
         }
-     return b;
-   }
- 
+        return b;
+    }
 }
