+++ week-01/Solidity-Speed-Quiz/src/RemoveElement.sol
@@ -2,7 +2,6 @@
 pragma solidity 0.8.28;
 
 contract RemoveElement {
-
     uint256[] public arr;
 
     function setArr(uint256[] memory _arr) public {
@@ -13,10 +12,9 @@ contract RemoveElement {
         // remove the item at index from arr
         // the new array should have a length of arr.length - 1
         // use swap-and-pop to remove the item
-        for(uint i=index; i<arr.length; i++){
-          arr[i] = arr[i+1];
+        for (uint256 i = index; i < arr.length; i++) {
+            arr[i] = arr[i + 1];
         }
         arr.pop();
-
     }
 }
